requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
django-restframework==0.0.1
djangoadmin==2.2.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangotools==2.2.2
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
pyasn1==0.4.8
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
rsa==4.8
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
